+++ src/main/java/kr/co/springbootweeklymission/domain/member/application/MemberService.java
@@ -1,13 +1,12 @@
 package kr.co.springbootweeklymission.domain.member.application;
 
 import kr.co.springbootweeklymission.domain.member.dao.MemberRepository;
-import kr.co.springbootweeklymission.domain.member.dto.response.MemberResDTO;
+import kr.co.springbootweeklymission.domain.member.api.response.MemberResDTO;
 import kr.co.springbootweeklymission.domain.member.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -15,10 +14,10 @@ public class MemberService {
     private final MemberRepository memberRepository;
 
     public List<MemberResDTO.READ> getMembersByBlack() {
-        List<Member> members = memberRepository.findMembersByBlack();
+        final List<Member> members = memberRepository.findMembersByBlack();
 
         return members.stream()
                 .map(Member::toMemberReadDto)
-                .collect(Collectors.toList());
+                .toList();
     }
 }
