+++ src/test/java/kr/co/springbootweeklymission/domain/voucher/entity/VoucherTest.java
@@ -1,6 +1,6 @@
 package kr.co.springbootweeklymission.domain.voucher.entity;
 
-import kr.co.springbootweeklymission.domain.model.VoucherPolicy;
+import kr.co.springbootweeklymission.domain.voucher.entity.model.VoucherPolicy;
 import kr.co.springbootweeklymission.domain.voucher.exception.WrongVoucherPolicyException;
 import kr.co.springbootweeklymission.global.error.model.ResponseStatus;
 import org.junit.jupiter.api.DisplayName;
